perf: simplify number_crossings function
This improved the performance of this function by ~60% according to the benchmarks

+++ analysis/src/timbral.rs
@@ -237,7 +237,7 @@ impl Normalize for SpectralDesc {
  */
 #[derive(Default, Clone)]
 pub struct ZeroCrossingRateDesc {
-    crossings_sum: u32,
+    crossings_sum: usize,
     samples_checked: usize,
 }
